Convert u16s to u32s (#1160)
* Convert u16s to u32s

* update test data

* fix: improve error message in get_next_sync_heights for better debugging

* update elfs

* update consts

* update test data

---------

+++ core/src/operator.rs
@@ -1509,7 +1509,7 @@ where
         }
 
         let watchtower_challenge_connector_start_idx =
-            (FIRST_FIVE_OUTPUTS + NUMBER_OF_ASSERT_TXS) as u16;
+            (FIRST_FIVE_OUTPUTS + NUMBER_OF_ASSERT_TXS) as u32;
 
         let bridge_circuit_host_params = BridgeCircuitHostParams::new_with_wt_tx(
             kickoff_tx.clone(),
